refactored conf.js to use utils functions
